Updated Custom directive templates code example
I have had a problem with the example code for the Custom directive templates. I was not able to bundle and minify this code with a .NET web optimization bundle. The error I was having:"Error: [$injector:unpr] Unknown provider: nProvider <- n <- mwlCalendarSlideBoxDirective ...". 

By adding the right dependencies for the '$provide' and '$delegate' the minification works again. I hope this helps someone in the future, at least my four hours spent finding a solution was worth something ;).