fix(weekViewWithTimes): call the cell-modifier for each days hour segment
BREAKING CHANGE:
The `cell-modifier` will now be called for every days hour segment instead of just the first day in the week.

The cssClass added will now be added on the segments day column instead of on the entire row.

The structure of the week view with times template has also changed slightly if using a custom template

Closes #424